Cleanup of MaterialX Format: - Moved export code to XmlExport.[cpp/h] - Removed some methods that didn't really belong in MaterialXFormat
