+++ source/MaterialXFormat/File.cpp
@@ -5,6 +5,8 @@
 
 #include <MaterialXFormat/File.h>
 
+#include <MaterialXFormat/Environ.h>
+
 #include <MaterialXCore/Exception.h>
 
 #if defined(_WIN32)
@@ -39,6 +41,7 @@ const string PATH_LIST_SEPARATOR = ";";
 #else
 const string PATH_LIST_SEPARATOR = ":";
 #endif
+const string MATERIALX_SEARCH_PATH_ENV_VAR = "MATERIALX_SEARCH_PATH";
 
 inline bool hasWindowsDriveSpecifier(const string& val)
 {
@@ -341,6 +344,12 @@ FilePath FilePath::getModulePath()
         }
     }
 #endif
+}
+
+FileSearchPath getEnvironmentPath(const string& sep)
+{
+    string searchPathEnv = getEnviron(MATERIALX_SEARCH_PATH_ENV_VAR);
+    return FileSearchPath(searchPathEnv, sep);
 }
 
 } // namespace MaterialX
