+++ source/MaterialXFormat/Util.cpp
@@ -5,44 +5,12 @@
 
 #include <MaterialXFormat/Util.h>
 
-#include <MaterialXFormat/Environ.h>
-
 #include <fstream>
 #include <iostream>
 #include <sstream>
 
 namespace MaterialX
 {
-
-namespace
-{
-
-FileSearchPath getSubDirectories(const FileSearchPath paths)
-{
-    std::set<std::string> childPathsSet;
-    FileSearchPath childPaths;
-    for (FileSearchPath::ConstIterator path = paths.begin(); path != paths.end(); ++path)
-    {
-        FilePath filePath = *path;
-        FilePathVec subdirs = filePath.getSubDirectories();
-        for (FilePath subdirPath : subdirs)
-        {
-            if (childPathsSet.count(subdirPath.asString()) == 0)
-            {
-                childPaths.append(subdirPath);
-                childPathsSet.emplace(subdirPath.asString());
-            }
-        }
-    }
-    return childPaths;
-}
-
-}
-
-FileSearchPath CORE_DEFINITION_PATH;
-const string MATERIALX_SEARCH_PATH_ENV_VAR = "MATERIALX_SEARCH_PATH";
-const string MATERIALX_ASSET_DEFINITION_PATH_ENV_VAR = "MATERIALX_ASSET_DEFINITION_PATH";
-const string MATERIALX_ASSET_TEXTURE_PATH_ENV_VAR = "MATERIALX_ASSET_TEXTURE_PATH";
 
 string readFile(const FilePath& filePath)
 {
@@ -172,32 +140,6 @@ StringSet loadLibraries(const FilePathVec& libraryFolders,
     return loadedLibraries;
 }
 
-StringSet loadCoreLibraries(const FilePathVec& libraryFolders,
-                            const FileSearchPath& searchPath,
-                            DocumentPtr doc,
-                            const StringSet& excludeFiles,
-                            XmlReadOptions* readOptions)
-{
-    FileSearchPath coreDefinitionPaths = getCoreDefinitionPath();
-    StringSet coreDefinitionPathsSet;
-    FilePathVec newCoreDefinitionPaths;
-    for (const FilePath& coreDefinitionPath : coreDefinitionPaths)
-    {
-        coreDefinitionPathsSet.emplace(coreDefinitionPath.asString());
-    }
-    for (const FilePath& libraryFolder : libraryFolders)
-    {
-        if (coreDefinitionPathsSet.count(libraryFolder.asString()) == 0)
-        {
-            newCoreDefinitionPaths.push_back(libraryFolder);
-            coreDefinitionPaths.append(libraryFolder);
-            coreDefinitionPathsSet.emplace(libraryFolder.asString());
-        }
-    }
-    setCoreDefinitionPath(coreDefinitionPaths);
-    return loadLibraries(newCoreDefinitionPaths, searchPath, doc, excludeFiles, readOptions);
-}
-
 void flattenFilenames(DocumentPtr doc, const FileSearchPath& searchPath, StringResolverPtr customResolver)
 {
     for (ElementPtr elem : doc->traverseTree())
@@ -257,87 +199,6 @@ void flattenFilenames(DocumentPtr doc, const FileSearchPath& searchPath, StringR
             elem->removeAttribute(Element::FILE_PREFIX_ATTRIBUTE);
         }
     }
-}
-
-FileSearchPath getResolvedDefinitionPath(const FileSearchPath& userDefinitionPath, bool includeSubFolders)
-{
-    const FileSearchPath coreDefinitionPath = getCoreDefinitionPath();
-    const FileSearchPath environmentPath = getEnvironmentPath();
-    const FileSearchPath assetDefinitionPath = getAssetDefinitionPath();
-
-    FileSearchPath resolvedDefinitionPath;
-    resolvedDefinitionPath.append(coreDefinitionPath);
-    resolvedDefinitionPath.append(environmentPath);
-    resolvedDefinitionPath.append(userDefinitionPath);
-    resolvedDefinitionPath.append(assetDefinitionPath);
-    if (includeSubFolders)
-    {
-        return getSubDirectories(resolvedDefinitionPath);
-    }
-    else
-    {
-        return resolvedDefinitionPath;
-    }
-}
-
-FileSearchPath getResolvedTexturePath(const FileSearchPath& userTexturePath, const FileSearchPath& userDefinitionPath, bool includeSubFolders)
-{
-    const FileSearchPath assetTexturePath = getAssetTexturePath();
-    const FileSearchPath resolvedDefinitionPath = getResolvedDefinitionPath(userDefinitionPath, includeSubFolders);
-
-    FileSearchPath resolvedTexturePath;
-    resolvedTexturePath.append(userTexturePath);
-    resolvedTexturePath.append(assetTexturePath);
-    if (includeSubFolders)
-    {
-        resolvedTexturePath = getSubDirectories(resolvedTexturePath);
-    }
-    resolvedTexturePath.append(resolvedDefinitionPath);
-
-    return resolvedTexturePath;
-}
-
-FileSearchPath getEnvironmentPath(const string& sep)
-{
-    string searchPathEnv = getEnviron(MATERIALX_SEARCH_PATH_ENV_VAR);
-    return FileSearchPath(searchPathEnv, sep);
-}
-
-FileSearchPath getAssetDefinitionPath(const string& sep)
-{
-    string assetDefinitionPathEnv = getEnviron(MATERIALX_ASSET_DEFINITION_PATH_ENV_VAR);
-    return FileSearchPath(assetDefinitionPathEnv, sep);
-}
-
-FileSearchPath getAssetTexturePath(const string& sep)
-{
-    string assetTexturePathEnv = getEnviron(MATERIALX_ASSET_TEXTURE_PATH_ENV_VAR);
-    return FileSearchPath(assetTexturePathEnv, sep);
-}
-
-FileSearchPath getCoreDefinitionPath()
-{
-    return CORE_DEFINITION_PATH;
-}
-
-void setEnvironmentPath(const FileSearchPath& path)
-{
-    setEnviron(MATERIALX_SEARCH_PATH_ENV_VAR, path.asString());
-}
-
-void setAssetDefinitionPath(const FileSearchPath& path)
-{
-    setEnviron(MATERIALX_ASSET_DEFINITION_PATH_ENV_VAR, path.asString());
-}
-
-void setAssetTexturePath(const FileSearchPath& path)
-{
-    setEnviron(MATERIALX_ASSET_TEXTURE_PATH_ENV_VAR, path.asString());
-}
-
-void setCoreDefinitionPath(const FileSearchPath& path)
-{
-    CORE_DEFINITION_PATH = path;
 }
 
 } // namespace MaterialX
