+++ source/MaterialXFormat/Util.h
@@ -20,11 +20,6 @@
 namespace MaterialX
 {
 
-extern MX_FORMAT_API const string PATH_LIST_SEPARATOR;
-extern MX_FORMAT_API const string MATERIALX_SEARCH_PATH_ENV_VAR;
-extern MX_FORMAT_API const string MATERIALX_ASSET_DEFINITION_PATH_ENV_VAR;
-extern MX_FORMAT_API const string MATERIALX_ASSET_TEXTURE_PATH_ENV_VAR;
-
 /// Read the given file and return a string containing its contents; if the read is not
 /// successful, then the empty string is returned.
 MX_FORMAT_API string readFile(const FilePath& file);
@@ -56,32 +51,12 @@ MX_FORMAT_API StringSet loadLibraries(const FilePathVec& libraryFolders,
                         const StringSet& excludeFiles = StringSet(),
                         const XmlReadOptions* readOptions = nullptr);
 
-/// Load the core library files within the given library folders into a document,
-/// using the given search path to locate the folders on the file system.
-MX_FORMAT_API StringSet loadCoreLibraries(const FilePathVec& libraryFolders,
-                        const FileSearchPath& searchPath,
-                        DocumentPtr doc,
-                        const StringSet& excludeFiles = StringSet(),
-                        XmlReadOptions* readOptions = nullptr);
-
 /// Flatten all filenames in the given document, applying string resolvers at the
 /// scope of each element and removing all fileprefix attributes.
 /// @param doc The document to modify.
 /// @param searchPath An optional search path for relative to absolute path conversion.
 /// @param customResolver An optional custom resolver to apply.
 MX_FORMAT_API void flattenFilenames(DocumentPtr doc, const FileSearchPath& searchPath = FileSearchPath(), StringResolverPtr customResolver = nullptr);
-
-MX_FORMAT_API FileSearchPath getResolvedDefinitionPath(const FileSearchPath& userDefinitionPath, bool includeSubFolders=true);
-MX_FORMAT_API FileSearchPath getResolvedTexturePath(const FileSearchPath& userTexturePath, const FileSearchPath& userDefinitionPath, bool includeSubFolders=true);
-
-MX_FORMAT_API FileSearchPath getEnvironmentPath(const string& sep = PATH_LIST_SEPARATOR);
-MX_FORMAT_API FileSearchPath getAssetDefinitionPath(const string& sep = PATH_LIST_SEPARATOR);
-MX_FORMAT_API FileSearchPath getAssetTexturePath(const string& sep = PATH_LIST_SEPARATOR);
-MX_FORMAT_API FileSearchPath getCoreDefinitionPath();
-MX_FORMAT_API void setEnvironmentPath(const FileSearchPath& path);
-MX_FORMAT_API void setAssetDefinitionPath(const FileSearchPath& path);
-MX_FORMAT_API void setAssetTexturePath(const FileSearchPath& path);
-MX_FORMAT_API void setCoreDefinitionPath(const FileSearchPath& path);
 
 } // namespace MaterialX
