+++ source/MaterialXFormat/XmlIo.cpp
@@ -266,14 +266,6 @@ unsigned int getParseOptions(const XmlReadOptions* readOptions)
     return parseOptions;
 }
 
-void mergeLooks(DocumentPtr doc, const XmlExportOptions* exportOptions)
-{
-    if (exportOptions && exportOptions->mergeLooks)
-    {
-        doc->mergeLooks(exportOptions->lookGroupToMerge);
-    }
-}
-
 } // anonymous namespace
 
 //
@@ -292,16 +284,6 @@ XmlReadOptions::XmlReadOptions() :
 
 XmlWriteOptions::XmlWriteOptions() :
     writeXIncludeEnable(true)
-{
-}
-
-//
-// XmlExportOptions methods
-//
-
-XmlExportOptions::XmlExportOptions() :
-    XmlWriteOptions(),
-    mergeLooks(false)
 {
 }
 
@@ -379,39 +361,6 @@ string writeToXmlString(DocumentPtr doc, const XmlWriteOptions* writeOptions)
     std::ostringstream stream;
     writeToXmlStream(doc, stream, writeOptions);
     return stream.str();
-}
-
-void exportToXmlStream(DocumentPtr doc, std::ostream& stream, const XmlExportOptions* exportOptions)
-{
-    mergeLooks(doc, exportOptions);
-    if (exportOptions && exportOptions->flattenFilenames)
-    {
-        FileSearchPath texturePath = getResolvedTexturePath(exportOptions->userTexturePath, exportOptions->userDefinitionPath);
-        flattenFilenames(doc, texturePath, exportOptions->stringResolver);
-    }
-    writeToXmlStream(doc, stream, exportOptions);
-}
-
-void exportToXmlFile(DocumentPtr doc, const FilePath& filename, const XmlExportOptions* exportOptions)
-{
-    mergeLooks(doc, exportOptions);
-    if (exportOptions && exportOptions->flattenFilenames)
-    {
-        FileSearchPath texturePath = getResolvedTexturePath(exportOptions->userTexturePath, exportOptions->userDefinitionPath);
-        flattenFilenames(doc, texturePath, exportOptions->stringResolver);
-    }
-    writeToXmlFile(doc, filename, exportOptions);
-}
-
-string exportToXmlString(DocumentPtr doc, const XmlExportOptions* exportOptions)
-{
-    mergeLooks(doc, exportOptions);
-    if (exportOptions && exportOptions->flattenFilenames)
-    {
-        FileSearchPath texturePath = getResolvedTexturePath(exportOptions->userTexturePath, exportOptions->userDefinitionPath);
-        flattenFilenames(doc, texturePath, exportOptions->stringResolver);
-    }
-    return writeToXmlString(doc, exportOptions);
 }
 
 void prependXInclude(DocumentPtr doc, const FilePath& filename)
